[FUSE] Handle SIGINT/SIGTERM to abort pending reads
Before, we would wait endlessly if there is a pending
read request that doesn't get answered (e.g. because
the server went down. That means you couldn't exit
the client in that case. Now we use a signal handler
to set a flag which causes the read to bail out and
return EIO.