[FUSE] Add new load balancing mechanism
The new mechanism is supposed to complement the existing
RTT based balancing. While the RTT-averaging approach is
better suited to react to sudden drastic changes in server
latency.
The new approach doesn't directly consider RTT, but keeps
track of how many consecutive times each server was the best
server when measuring the RTTs. The higher that counter
rises, the more likely it will become that the connection
switches over to that server.
Eg.: Server 1 measures 600µs each time
     Server 2 measures 599µs each time
After a while, in case server 1 is currently used, the
connection will eventually switch over to server 2. The
RTT-based mechanism would not switch over in this case,
since the threshold that prevents constant switching between
servers is not reached.

The new approach is meant to handle scenarios where the
network is generally fast, but it would still be beneficial
from a network topology point of view if the clients switch
to the slightly faster server, assuming it is closer to the
client and thus less network segments are burdened.