[FUSE] In panic mode, use a pending range for alt check
If we lost connection and then go check all known alt
servers, see if we have some pending request queued and
if so, use its offset and length for the alt server probe.
This ensures that the server being tested is able to
satisfy at least the next request we'll send.