[KERNEL] convert to blk-mq and ktime
This converts the dnbd3 kernel module driver to use the blk-mq
infrastructure,  which allows the dnbd3 kernel module driver to be
compatible with Linux kernels in version 5.x or later. The conversion of
the implementation uses one hardware queue to preserve the existing
send/receive and load-balancing logic, but can be scaled up in the
future. In addition to that, time measurements in the implementation are
converted to ktime based accessors to replace the use of deprecated time
interfaces.