[KERNEL] set socket timeouts with new function and time structure
This patch replaces the deprecated kernel_setsockopt() function with the
socket generic function sock_setsockopt() to set the send and receive
timeout for all used sockets in the kernel module. The timeout data type
is choosen according to the available kernel version, so that the dnbd3
kernel module supports the new y2038-problem-solving socket options.
With both changes, the kernel module runs on Linux kernel 5.4 LTS and
later (including Linux kernel 5.9).