[SERVER] Update nextSave timestamp at start of function
saveLoadAllCacheMaps() is called frequently, and a 'full' run can take
some time. If we only update the nextSave timestamp when we're done, we
might already have a concurrent call to the function, which will also do
a 'full' run, since the timestamp is not updated yet. This doesn't break
anything, but leads to even more disk activity, which is probably
already high, given that the previous run is not done yet.