[BUILD] build docker containers only if build of dnbd3-server is enabled
This patch fixes the issue that a Docker container for the dnbd3
components (including the dnbd3-server) can be built with missing
dnbd3-server's binary. Since the created Docker container starts
the missing binary, it will fail. Therefore, a built of the Docker
image is only possible if building of the dnbd3-server is enabled
(CMAKE option DNBD3_SERVER=ON).