[BUILD] add support for atomic operations not supported by hardware
This change links the dnbd3-server with 'libatomic' to add support for
atomic operations not supported by hardware (especially 32-bit hardware
architectures, such as ARM). Thus the dnbd3-server can also run on a
Raspberry Pi 1 running Rasperry Pi OS.

Note that the dnbd3-server is still linked to the libatomic, even if the
hardware supports atomic operations. In this case, the compiler resolves
atomic operations and replaces them automatically with specific built-in
functions. This unnecessary linkage can be removed in the future if the
GCC supports an upcoming option called automatic linking of libatomic
(--enable-autolink-libatomic).