[SERVER] Mark uplink requests with BGR/prefetch flags and handle them
Incoming requests from clients might actually be prefetch jobs from
another downstream proxy. Don't do prefetching for those, as this would
cascade upwars in the proxy chain (prefetch for a prefetch of a prefetch)

Incoming requests might also be background replication. Don't relay
those if we're not configured for background replication as well.