[BUILD] CheckPatch: Quote variable to avoid error if empty
The variable is most likely empty if the according kernel headers
directory was deleted. However, as it currently stands, cmake is
complaining about too few arguments, which isn't really helpful to the
user.
Instead, quote the variable to make cmake finish, and have the actual
compile process of the kernel module fail, which will yield a much more
helpful error message.