[FUSE] Make initial connect entirely parallel wrt servers
Previously, we only did the actual socket connect in a
concurrent matter. Once a connection was successfully
established, we did a blocking handshake on the protocol
level. If the server war particularly slow, this was bad
as we would not try other servers until after this.
Throw out the previous non-blocking async connect logic,
switching to a multi-threaded approach, that spawns one
thread per host/ip, offset by 200ms, until one of
the attempts succeeds.